rename deltabot.py to bot.py, make DeltaBot object available top-level
