+++ tox.ini
@@ -28,6 +28,7 @@ commands = check-manifest
 
 
 [pytest]
+addopts = -v 
 filterwarnings =
     ignore::DeprecationWarning
 python_files = tests/test*.py src/deltabot/builtin/*.py examples/*.py
